Refactor and create InstallApplication
